requirements.txt
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
decord==0.6.0
einops==0.8.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.34.1
idna==3.10
imageio==2.37.0
inflate64==1.0.3
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pip==25.1
psutil==7.0.0
py7zr==1.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
pyparsing==3.2.3
pyppmd==1.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzstd==0.17.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.15.3
setuptools==78.1.1
six==1.17.0
scikit-learn==1.7.1
smplx==0.1.28
submitit==1.5.3
sympy==1.14.0
texttable==1.7.0
timm==1.0.19
tokenizers==0.22.1
torch==2.7.1
torchprofile==0.0.4
torchvision==0.22.1
tqdm==4.67.1
transformers==4.56.2
triton==3.3.1
typing_extensions==4.14.1
urllib3==2.5.0
uv==0.8.8
wandb==0.21.1
wheel==0.45.1
xformers==0.0.31.post1
